MGMT-6699 - Move libvirt functions to controller. Move unused utils functions to deprecated_utils.py.
