+++ discovery-infra/day2.py
@@ -6,6 +6,9 @@ import time
 import uuid
 
 import waiting
+
+import test_infra.utils.waiting
+from deprecated_utils import wait_till_nodes_are_ready, are_libvirt_nodes_in_cluster_hosts
 from test_infra import assisted_service_api, utils, consts, warn_deprecate
 from test_infra.tools import static_network, terraform_utils
 
@@ -110,12 +113,12 @@ def day2_nodes_flow(client,
         num_nodes_to_wait = num_worker_nodes
         installed_status = consts.NodesStatus.DAY2_INSTALLED
 
-    utils.wait_till_nodes_are_ready(
+    wait_till_nodes_are_ready(
         nodes_count=num_nodes_to_wait, network_name=tf_network_name
     )
 
     waiting.wait(
-        lambda: utils.are_libvirt_nodes_in_cluster_hosts(
+        lambda: are_libvirt_nodes_in_cluster_hosts(
             client, cluster.id, num_nodes_to_wait
         ),
         timeout_seconds=consts.NODES_REGISTERED_TIMEOUT,
@@ -130,7 +133,7 @@ def day2_nodes_flow(client,
                                   tf_network_name,
                                   cluster.id)
 
-    utils.wait_till_all_hosts_are_in_status(
+    test_infra.utils.waiting.wait_till_all_hosts_are_in_status(
         client=client,
         cluster_id=cluster.id,
         nodes_count=num_worker_nodes,
@@ -150,7 +153,7 @@ def day2_nodes_flow(client,
         log.info(
             "Start waiting until all nodes of cluster %s have been installed( reached added-to-existing-clustertate)",
             cluster.id)
-        utils.wait_till_all_hosts_are_in_status(
+        test_infra.utils.waiting.wait_till_all_hosts_are_in_status(
             client=client,
             cluster_id=cluster.id,
             nodes_count=num_nodes_to_wait,
