+++ discovery-infra/test_infra/consts/consts.py
@@ -27,6 +27,7 @@ HOST_PASSTHROUGH_CPU_MODE = "host-passthrough"
 MASTER_TF_CPU_MODE = HOST_PASSTHROUGH_CPU_MODE
 WORKER_TF_CPU_MODE = HOST_PASSTHROUGH_CPU_MODE
 NODES_REGISTERED_TIMEOUT = 20 * MINUTE
+DEFAULT_CHECK_STATUSES_INTERVAL = 5
 CLUSTER_INSTALLATION_TIMEOUT = HOUR
 CLUSTER_INSTALLATION_TIMEOUT_OCS = 95 * MINUTE
 START_CLUSTER_INSTALLATION_TIMEOUT = 6 * MINUTE
